Clarify README.md
Clarify that this repo is a component of a murder project and not a project template. 

hi, i ended up fumbling around for a bit trying to make some documentation for myself before i realized that this was not a project template, just wanted to clarify and update others so that this mistake isn't repeated lol.