Switch from MonoGame to FNA (#78)
Issue
I was more comfortable working with FNA to get the latest Native AOT updates. I was not very happy with the lack of nuget packages, which I workarounded by maintaining the feed myself for the engine at murder-fna.

Changes
I felt that a PR would be a more organized way to track this work, if we ever want to switch back. In any case, I felt that there were lots of related cleanups that were good for the overall architecture, so we'll keep that anyway.

---------

* Ongoing changes.

* Remove previous shaders. Come up with a solution to find fxc.exe installed...

* Enable optional shaders.

* Cleanup cursor logic.

* Try to improve SDL API for mouse.

* Fix text input for FNA.

* major shaders overhaul

* Add murder shaders.

* Use nuget package (hehe).

* Update nuget.

* Bump FNA and other bindings.

---------

Co-authored-by: saint11 <saintjust@gmail.com>